requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorcet==3.1.0
contourpy==1.1.1
cycler==0.12.1
dataclasses-json==0.6.6
distro==1.9.0
Django==4.2.8
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
Flask==3.0.0
Flask-Compress==1.15
Flask-Cors==4.0.1
Flask-Sockets==0.2.1
fonttools==4.51.0
frozenlist==1.4.1
gevent==24.2.1
gevent-websocket==0.10.1
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
groq==0.9.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
imageio==2.34.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-text-splitters==0.0.1
langsmith==0.1.57
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.2
matplotlib==3.7.5
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.4
openai==1.12.0
opencv-contrib-python==4.9.0.80
opencv-python==4.6.0.66
orjson==3.10.3
packaging==23.2
pandas==1.5.1
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
protobuf==4.25.3
psycopg2==2.9.9
pyarrow==15.0.0
pycparser==2.22
pydantic==2.6.2
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
runway-python==0.6.1
scikit-image==0.17.2
scikit-learn==1.1.3
scipy==1.10.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.30
sqlparse==0.4.4
starlette==0.37.2
streamlit==1.31.1
tenacity==8.2.3
threadpoolctl==3.5.0
tifffile==2023.7.10
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
Unidecode==1.3.8
urllib3==2.2.1
uvicorn==0.30.3
validators==0.22.0
watchdog==4.0.0
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
zope.event==5.0
zope.interface==6.4.post2
zstandard==0.23.0
